Switch to a subset of ts-standard from Prettier
We've long valued Prettier's configurability and convenience,
but we were somehow unaware
of the maintainer's aggressive and
demeaning stance toward
devs who actually want their code
to look beautiful to *them*.
They now refuse to ever implement
new configuration options
and have expressed remorese over
adding them in the first place.

Given our sensory differences,
we find that semicolons
make code harder to read.
So we have always disabled semicolons
when using Prettier, in contradiction
of the default for TypeScript
(according to them).

ts-standard (and StandardJS), by contrast, have defaults that are *much*
more in line with our idea of
aesthetically pleasing code...
And Standard is deliberately implemented
as an ESLint shared config,
which means rules can be overridden
with ease by design.

And that's just it, at the end of the day:
The features of Prettier were ultimately
purely cosmetic. And if the code is not aesthetically
pleasing ---- hell, if it's bloody unreadable ---- to its own devs,
what good is a style guide?

TL;DR Style guides are great.
Attempting to force devs to follow them in contradiction of their
own sense of aesthetics is awful.